moved table in dialog to separate file which is reused by all dialogs
