+++ src/modules/Preservation/views/ScopeOverview/DialgTable.style.ts
