+++ src/modules/Preservation/views/ScopeOverview/StartPreservationDialog.tsx
@@ -2,70 +2,28 @@ import React from 'react';
 import { PreservedTag } from './types';
 import { Typography } from '@equinor/eds-core-react';
 import { tokens } from '@equinor/eds-tokens';
-import Table from '../../../../components/Table';
 import RequirementIcons from './RequirementIcons';
-import { Toolbar } from './StartPreservationDialog.style';
-import { isTagOverdue } from './ScopeOverview';
+import DialogTable from './DialogTable';
+import { Column } from 'material-table';
 
 interface StartPreservationDialogProps {
     startableTags: PreservedTag[];
     nonStartableTags: PreservedTag[];
 }
 
-interface TableProps {
-    tags: PreservedTag[];
-    toolbarText: string;
-    toolbarColor: string;
-}
-
 const getRequirementIcons = (tag: PreservedTag): JSX.Element => {
+    console.log(tag);
     return (
         <RequirementIcons tag={tag} />
     );
 };
 
-const StartTable = ({
-    tags,
-    toolbarText,
-    toolbarColor
-}: TableProps): JSX.Element => {
-
-    const numTags = tags.length;
-
-    return (<Table
-        columns={[
-            { title: 'Tag nr', field: 'tagNo' },
-            { title: 'Description', field: 'description' },
-            { title: 'Status', field: 'status' },
-            { title: 'Req type', render: getRequirementIcons }
-        ]}
-        data={tags}
-        options={{
-            search: false,
-            pageSize: 5,
-            pageSizeOptions: [5, 10, 50, 100],
-            showTitle: false,
-            draggable: false,
-            selection: false,
-            emptyRowsWhenPaging: false,
-            headerStyle: {
-                backgroundColor: tokens.colors.interactive.table__header__fill_resting.rgba
-            },
-            rowStyle: (rowData): any => ({
-                color: isTagOverdue(rowData) && tokens.colors.interactive.danger__text.rgba,
-            }),
-        }}
-        components={{
-            Toolbar: (): any => (
-                <Toolbar>
-                    <Typography style={{ color: toolbarColor }} variant='h6' >{numTags} {toolbarText}</Typography>
-                </Toolbar>
-            )
-        }}
-
-        style={{ boxShadow: 'none' }}
-    />);
-};
+const columns: Column<any>[] = [
+    { title: 'Tag nr', field: 'tagNo' },
+    { title: 'Description', field: 'description' },
+    { title: 'Status', field: 'status' },
+    { title: 'Req type', render: getRequirementIcons }
+];
 
 const StartPreservationDialog = ({
     startableTags,
@@ -76,11 +34,11 @@ const StartPreservationDialog = ({
         {nonStartableTags.length > 0 && (
             <div>
                 <Typography variant="meta">{nonStartableTags.length} tag(s) cannot be started.</Typography>
-                <StartTable tags={nonStartableTags} toolbarText='tag(s) will not be started' toolbarColor={tokens.colors.interactive.danger__text.rgba} />
+                <DialogTable tags={nonStartableTags} columns={columns} toolbarText='tag(s) will not be started' toolbarColor={tokens.colors.interactive.danger__text.rgba} />
             </div>
         )}
         {startableTags.length > 0 && (
-            <StartTable tags={startableTags} toolbarText='tag(s) will be started' toolbarColor={tokens.colors.interactive.primary__resting.rgba} />
+            <DialogTable tags={startableTags} columns={columns} toolbarText='tag(s) will be started' toolbarColor={tokens.colors.interactive.primary__resting.rgba} />
         )}
     </div>
     );
