--- src/modules/Preservation/views/ScopeOverview/TransferDialog.style.ts
@@ -1,6 +1,0 @@
-import styled from 'styled-components';
-
-export const Toolbar = styled.div`
-    margin-top: calc(var(--grid-unit) * 2);
-    margin-bottom: var(--grid-unit);
-`;
